+++ .gitignore
@@ -1,5 +1,6 @@
 *.pyc
 *.egg-info/
+*.swp
 deps/
 include/
 lib/
