+++ tools/c7n_mailer/c7n_mailer/handle.py
@@ -17,20 +17,13 @@ Lambda entry point
 
 import base64
 import boto3
-import getpass
 import json
-import logging
 import os
 
-
-logging.root.setLevel(logging.DEBUG)
-logging.getLogger('botocore').setLevel(logging.WARNING)
-
-log = logging.getLogger('custodian.mailer')
+from .sqs_queue_processor import MailerSqsQueueProcessor
 
 
-def bootstrap():
-    log.debug("Initializing")
+def config_setup(session):
     task_dir = os.environ.get('LAMBDA_TASK_ROOT')
     os.environ['PYTHONPATH'] = "%s:%s" % (task_dir, os.environ.get('PYTHONPATH', ''))
     with open(os.path.join(task_dir, 'config.json')) as fh:
@@ -46,24 +39,12 @@ def bootstrap():
         os.environ['https_proxy'] = config['https_proxy']
     return config
 
-session = boto3.Session()
-config = bootstrap()
 
-
-def run(event, context):
+def start_c7n_mailer(event, context, logger):
     try:
-        from markupsafe import Markup
-        from jinja2 import utils
-        from .worker import Worker
-        from .processor import Processor
+        session = boto3.Session()
+        config = config_setup(session)
+        mailer_sqs_queue_processor = MailerSqsQueueProcessor(config, session, logger)
+        mailer_sqs_queue_processor.run()
     except Exception as e:
-        log.exception("import failed %s", e)
-
-    try:
-        log.info("Worker Run")
-        w = Worker(config, context, session)
-        w.run()
-    except:
-        log.exception("Error processing worker \n DebugEnv: %s \n User: %s \n" % (
-            os.environ['PYTHONPATH'],
-            getpass.getuser()))
+        logger.exception("Error starting mailer MailerSqsQueueProcessor(). \n Error: %s \n" % (e))
