+++ tox.ini
@@ -1,0 +1,10 @@
+[flake8]
+exclude = .git,venv,python2.7,bin,test_*
+; E401 is 'multiple imports on one line' -- no big deal
+; E128 is under identation
+; E203 is whitespace before ':'
+; E221 is multiple spaces before operator
+; E231 is missing whitespace after  ,/:
+; E251 is unexpected spaces around keywords
+ignore = E401,E128,E203,E221,E231,E251
+max-line-length = 100
